project-root4 - Copy/document-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
boto3==1.29.7
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
minio==7.2.0
